build error handled
