+++ src/app/register/viewport.ts
@@ -1,0 +1,6 @@
+export const viewport = {
+  width: 'device-width',
+  initialScale: 1,
+  maximumScale: 1,
+  userScalable: false,
+};
